% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_models.R
\name{cohens_d_analogue}
\alias{cohens_d_analogue}
\title{Calculate Cohen's d analogue for mixed effects models}
\usage{
cohens_d_analogue(glht, model)
}
\arguments{
\item{glht}{General linear hypothesis test object}

\item{model}{Mixed effects model}
}
\value{
Matrix of Cohen's d analogue values
}
\description{
The ratio of mean to variance is used as a measure of effect size, with the
combined residual variance and random effect variance in the denominator
(analogous to Cohen's D but for LME models).
}
