% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_models.R
\name{cohens_local_fsquared}
\alias{cohens_local_fsquared}
\title{Calculate Cohen's local f-squared}
\usage{
cohens_local_fsquared(model_a, model_b)
}
\arguments{
\item{model_a}{The model with the effect of interest included}

\item{model_b}{The model with the effect of interest removed}
}
\value{
Cohen's f-squared value
}
\description{
For estimation statistics, Cohen's f2 (f-squared) is a measure of effect
size. The local version of the f2 statistic quantifies the amount of variance
accounted for by variables of interest, relative to the amount of variance
left unaccounted for by the model.
}
