% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{correlation_interaction_matrix}
\alias{correlation_interaction_matrix}
\title{Test for significant correlation between a continuous response variable and
a continuous covariate, in the presence of an interaction between experiment
and covariate.}
\format{
Matrix (3x6) with experiment contrasts: Novelty, Locomotion, Dual
}
\usage{
correlation_interaction_matrix
}
\description{
Test for significant correlation between a continuous response variable and
a continuous covariate, in the presence of an interaction between experiment
and covariate.
}
\keyword{datasets}
