% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{day_occupancy_contrast_matrix}
\alias{day_occupancy_contrast_matrix}
\title{Tests for significant differences in occupancy between COW and NOD, by
sleep state.}
\format{
Matrix (5x10) with COW-NOD contrasts: NREM, IS, REM, Wake, MA
}
\usage{
day_occupancy_contrast_matrix
}
\description{
Tests for significant differences in occupancy between COW and NOD, by
sleep state.
}
\keyword{datasets}
