% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ewk_change_dual_matrix}
\alias{ewk_change_dual_matrix}
\title{Tests for extended wake change in dual experiments, in the presence of an
interction between experiment and condition.}
\format{
Matrix (1x6) with contrast: L.EWK-E.EWK
}
\usage{
ewk_change_dual_matrix
}
\description{
Tests for extended wake change in dual experiments, in the presence of an
interction between experiment and condition.
}
\keyword{datasets}
