% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ewk_change_main_effect_matrix}
\alias{ewk_change_main_effect_matrix}
\title{Tests for extended wake change when there is no interaction between
experiment and condition, and a significant main effect of condition.}
\format{
Matrix (1x4) with contrast: L.EWK-E.EWK
}
\usage{
ewk_change_main_effect_matrix
}
\description{
Tests for extended wake change when there is no interaction between
experiment and condition, and a significant main effect of condition.
}
\keyword{datasets}
