% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{exp_occupancy_contrast_matrix}
\alias{exp_occupancy_contrast_matrix}
\title{Tests for significant differences in occupancy between REC and BSL, by
sleep state.}
\format{
Matrix (5x10) with REC-BSL contrasts: NREM, IS, REM, Wake, MA
}
\usage{
exp_occupancy_contrast_matrix
}
\description{
Tests for significant differences in occupancy between REC and BSL, by
sleep state.
}
\keyword{datasets}
