% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_shapes.R
\docType{data}
\name{experiment_colors}
\alias{experiment_colors}
\title{Custom colors for experiments}
\format{
An object of class \code{character} of length 3.
}
\usage{
experiment_colors
}
\description{
Named vector of colors for different experimental conditions.
}
\keyword{datasets}
