% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_shapes.R
\docType{data}
\name{experiment_shapes}
\alias{experiment_shapes}
\title{Custom shapes for experiments}
\format{
An object of class \code{numeric} of length 3.
}
\usage{
experiment_shapes
}
\description{
Named vector of point shapes for different experimental conditions.
}
\keyword{datasets}
