% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_models.R
\name{fit_basic_models}
\alias{fit_basic_models}
\title{Fit basic statistical models}
\usage{
fit_basic_models(d, response_var, primary_covariate, random_factor)
}
\arguments{
\item{d}{Data frame}

\item{response_var}{Response variable name}

\item{primary_covariate}{Primary covariate name}

\item{random_factor}{Random factor name}
}
\value{
List of fitted models (full, null)
}
\description{
Fits basic full and null models for simpler comparisons.
}
