% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_testing.R
\name{fit_dummy_models_for_fsquared}
\alias{fit_dummy_models_for_fsquared}
\title{Fit dummy models for f-squared calculation}
\usage{
fit_dummy_models_for_fsquared(
  d,
  response_var,
  primary_covariate,
  secondary_covariate,
  random_factor
)
}
\arguments{
\item{d}{Data frame}

\item{response_var}{Response variable name}

\item{primary_covariate}{Primary covariate name}

\item{secondary_covariate}{Secondary covariate name}

\item{random_factor}{Random factor name}
}
\value{
List with dummy-coded data and models
}
\description{
If you have covariate X (hippocampal theta-delta ratio),
categorical variable Z (experiment) with levels A (NOD), B (COW), C (CTN).
}
\details{
The full model y ~ X * Z will have terms
\code{[intercept, zB, zC, X, zB:X, zC:X]}.
To test for an effect for A (Novelty), you test if \code{X = 0}.
To test for an effect for B (Locomotion), you test if \code{X + zB:X = 0}.
To test for an effect for C (Locomotion + Novelty), you test if
\code{X + zC:X = 0}.
}
