% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_tests.R
\name{get_anova_pval}
\alias{get_anova_pval}
\title{Statistical testing functions}
\usage{
get_anova_pval(anova_result)
}
\arguments{
\item{anova_result}{Result from anova() function}
}
\value{
P-value from likelihood ratio test
}
\description{
General functions for statistical hypothesis testing.
Get ANOVA p-value
}
