% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_testing.R
\name{get_correlation_data}
\alias{get_correlation_data}
\title{Correlation testing functions}
\usage{
get_correlation_data(d, x_condition, x_variable, y_condition, y_variable)
}
\arguments{
\item{d}{Data frame in long format}

\item{x_condition}{Condition name for x variable}

\item{x_variable}{Variable name for x}

\item{y_condition}{Condition name for y variable}

\item{y_variable}{Variable name for y}
}
\value{
List with merged data frame and variable names
}
\description{
Functions for correlation analyses between conditions and variables.
Get correlation data
}
\details{
Transform data for correlation analysis between two condition/variable pairs.
If x_condition = "EXT.Wake", x_variable = "Theta",
y_condition = "REC.NREM", y_variable = "SPWs"...

Transform data d of the form:

subject experiment condition THETA SPWs
1       A          EXT.Wake   10    20
1       A          REC.NREM   15    25
1       B          EXT.Wake   20    30
1       B          REC.NREM   25    35

into dxy of the form:

subject experiment \code{EXT.Wake.Theta} \code{REC.NREM.SPWs}
1       A          10              25
1       B          20              35
2       A          ...             ...

where xvar = "\code{EXT.Wake.Theta}" and yvar = "\code{REC.NREM.SPWs}"
}
