% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_testing.R
\name{get_fsquared_for_interaction_posthocs}
\alias{get_fsquared_for_interaction_posthocs}
\title{Get f-squared for interaction post-hocs}
\usage{
get_fsquared_for_interaction_posthocs(
  d,
  response_var,
  primary_covariate,
  secondary_covariate,
  random_factor
)
}
\arguments{
\item{d}{Data frame}

\item{response_var}{Response variable name}

\item{primary_covariate}{Primary covariate name}

\item{secondary_covariate}{Secondary covariate name}

\item{random_factor}{Random factor name}
}
\value{
List of effect sizes by level
}
\description{
Get Cohen's f-squared for posthoc tests, when Cohen's d is not appropriate.
The issue is that Cohen's d estimates may depend on the scale of the inputs.
You can change the scale of the inputs, making the estimates larger or
smaller arbitrarily, and the residual error won't change (linear
transformation of a linear model). So that notion of effect size doesn't
work, but something like Cohen's f-squared does.
}
