% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{homeostasis_dual_matrix}
\alias{homeostasis_dual_matrix}
\title{Tests for homeostatic effects in dual experiments, in the presence of an
interction between experiment and condition.}
\format{
Matrix (5x18) with contrasts:
E.REC-L.BSL, L.REC-E.REC, L.EWK-E.EWK, E.BSL-L.BSL, E.REC-E.BSL
}
\usage{
homeostasis_dual_matrix
}
\description{
Tests for homeostatic effects in dual experiments, in the presence of an
interction between experiment and condition.
}
\keyword{datasets}
