% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{homeostasis_main_effect_matrix}
\alias{homeostasis_main_effect_matrix}
\title{Tests for homeostatic effects when there is no interaction between experiment
and condition, and a significant main effect of condition.}
\format{
Matrix (5x8) with contrasts:
E.REC-L.BSL, L.REC-E.REC, L.EWK-E.EWK, E.BSL-L.BSL, E.REC-E.BSL
}
\usage{
homeostasis_main_effect_matrix
}
\description{
Tests for homeostatic effects when there is no interaction between experiment
and condition, and a significant main effect of condition.
}
\keyword{datasets}
