% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{homeostasis_novelty_matrix}
\alias{homeostasis_novelty_matrix}
\title{Contrast matrices for sleep homeostasis and extended wake testing
Tests for homeostatic effects in novelty experiments, in the presence of an
interction between experiment and condition.}
\format{
Matrix (5x18) with contrasts:
E.REC-L.BSL, L.REC-E.REC, L.EWK-E.EWK, E.BSL-L.BSL, E.REC-E.BSL
}
\usage{
homeostasis_novelty_matrix
}
\description{
Contrast matrices for sleep homeostasis and extended wake testing
Tests for homeostatic effects in novelty experiments, in the presence of an
interction between experiment and condition.
}
\keyword{datasets}
