% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{load_condition_differences}
\alias{load_condition_differences}
\title{Load condition differences data}
\usage{
load_condition_differences(rev = FALSE)
}
\arguments{
\item{rev}{Logical, whether to load revised multicortical version}
}
\value{
Data frame with condition differences data
}
\description{
Load condition differences data
}
