% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{load_measures}
\alias{load_measures}
\title{Data loading functions}
\usage{
load_measures(rev = FALSE)
}
\arguments{
\item{rev}{Logical, whether to load revised multicortical version}
}
\value{
Data frame with measures data
}
\description{
Functions to load various datasets from the package extdata directory.
Load measures data
}
