% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{load_measures_for_ewk_change_testing}
\alias{load_measures_for_ewk_change_testing}
\title{Load measures for extended wake change testing}
\usage{
load_measures_for_ewk_change_testing()
}
\value{
Data frame filtered for extended wake conditions
}
\description{
Load measures for extended wake change testing
}
