% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{load_measures_for_first_novelty_testing}
\alias{load_measures_for_first_novelty_testing}
\title{Load measures for first novelty testing}
\usage{
load_measures_for_first_novelty_testing(only_dual_subjects)
}
\arguments{
\item{only_dual_subjects}{Logical, whether to include only dual experiment subjects}
}
\value{
Data frame filtered for first novelty conditions
}
\description{
Load measures for first novelty testing
}
