% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{load_measures_for_homeostasis_testing}
\alias{load_measures_for_homeostasis_testing}
\title{Load measures for homeostasis testing}
\usage{
load_measures_for_homeostasis_testing(rev = FALSE)
}
\arguments{
\item{rev}{Logical, whether to load revised multicortical version}
}
\value{
Data frame filtered for homeostasis conditions
}
\description{
Load measures for homeostasis testing
}
