% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{load_occupancy_data}
\alias{load_occupancy_data}
\title{Load occupancy data}
\usage{
load_occupancy_data()
}
\value{
Data frame with sleep period fractional occupancy data
}
\description{
Load occupancy data
}
