% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{load_rec_rem_vs_ewk_tdr_data}
\alias{load_rec_rem_vs_ewk_tdr_data}
\title{Load recovery REM vs extended wake TDR data}
\usage{
load_rec_rem_vs_ewk_tdr_data()
}
\value{
Data frame with recovery REM vs extended wake TDR data
}
\description{
Load recovery REM vs extended wake TDR data
}
