% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{load_ripple_frequency_data}
\alias{load_ripple_frequency_data}
\title{Load ripple frequency data}
\usage{
load_ripple_frequency_data()
}
\value{
Data frame with ripple frequency data
}
\description{
Load ripple frequency data
}
