% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{load_sgfg_by_epoch_type}
\alias{load_sgfg_by_epoch_type}
\title{Load SG/FG data by epoch type}
\usage{
load_sgfg_by_epoch_type()
}
\value{
Data frame with SG/FG data
}
\description{
Load SG/FG data by epoch type
}
