% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_correlation_without_experiment_interaction}
\alias{plot_correlation_without_experiment_interaction}
\title{Plot condition correlation without interaction}
\usage{
plot_correlation_without_experiment_interaction(
  result,
  response_var,
  primary_covariate,
  fig_dir = NULL
)
}
\arguments{
\item{result}{Result object from correlation test without significant
interaction}

\item{response_var}{Name of response variable (y-axis)}

\item{primary_covariate}{Name of primary covariate (x-axis)}

\item{fig_dir}{Directory to save figures to. If NULL, figures are not saved.}
}
\value{
ggplot2 plot object
}
\description{
Creates a plot showing correlation results when there is no significant
interaction with experiment.
}
