% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_models.R
\name{subtract_nested_random_effects}
\alias{subtract_nested_random_effects}
\title{Subtract nested random effects from a data frame}
\usage{
subtract_nested_random_effects(d, m)
}
\arguments{
\item{d}{The data with random effects included, used to fit model m}

\item{m}{The model with random effects structure that must be removed}
}
\value{
Data frame with random effects subtracted
}
\description{
Subtract nested random effects from a data frame
}
