% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_models.R
\name{subtract_random_effects}
\alias{subtract_random_effects}
\title{Subtract random effects from data}
\usage{
subtract_random_effects(d, m)
}
\arguments{
\item{d}{The data with random effects included, used to fit model m}

\item{m}{The model with random effects structure that must be removed}
}
\value{
Data frame with random effects subtracted
}
\description{
In order to accurately estimate the local effect size of an interaction,
we need to fit standard linear models that have the random effects
removed from the data.
}
