% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_models.R
\name{subtract_ranef_get_fsquared}
\alias{subtract_ranef_get_fsquared}
\title{Subtract random effects and calculate f-squared}
\usage{
subtract_ranef_get_fsquared(d, model_a, model_b)
}
\arguments{
\item{d}{Data frame}

\item{model_a}{The model with the effect of interest included, and the source
of random effects to be subtracted from the data}

\item{model_b}{The model with the effect of interest removed}
}
\value{
List with adjusted data, models, and f-squared value
}
\description{
There is a caveat to computing f2 for LME models; when we remove the effect
of interest, it is possible that the reduced model is estimated with a
drastically different random effect structure, misrepresenting the variance
accounted for by the fixed effects (Selya et al., 2012). To circumvent this
issue, we first fit the full model to estimate a random effect structure
and subtract the estimated random effects from the data. We then fit two
standard linear models to the adjusted data (one full, one reduced) and use
their R2 values to calculate f2. In this way, we enforce that the full and
reduced models have identical random effect structure, and that f2 explicitly
captures the variance accounted for by the fixed effects.
}
