% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_models.R
\name{subtract_simple_random_effects}
\alias{subtract_simple_random_effects}
\title{Subtract simple (non-crossed, non-nested, single-level) random effects}
\usage{
subtract_simple_random_effects(d, m)
}
\arguments{
\item{d}{The data with random effects included, used to fit model m}

\item{m}{The model with random effects structure that must be removed}
}
\value{
Data frame with random effects subtracted
}
\description{
Subtract simple (non-crossed, non-nested, single-level) random effects
}
