% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_testing.R
\name{test_condition_correlation}
\alias{test_condition_correlation}
\title{Test condition correlation}
\usage{
test_condition_correlation(x_condition, x_variable, y_condition, y_variable)
}
\arguments{
\item{x_condition}{Condition name for x variable}

\item{x_variable}{Variable name for x}

\item{y_condition}{Condition name for y variable}

\item{y_variable}{Variable name for y}
}
\value{
List with correlation test results
}
\description{
Test correlation between variables from different conditions.
}
