% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_testing.R
\name{test_condition_diff_corr}
\alias{test_condition_diff_corr}
\title{Test condition difference correlation}
\usage{
test_condition_diff_corr(x_variable, y_variable)
}
\arguments{
\item{x_variable}{Variable name for x}

\item{y_variable}{Variable name for y}
}
\value{
List with correlation test results
}
\description{
Test correlation between variables using condition difference data.
}
