% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_testing.R
\name{test_ewk_change}
\alias{test_ewk_change}
\title{Other specialized testing functions}
\usage{
test_ewk_change(response_variable)
}
\arguments{
\item{response_variable}{Name of the response variable to test}
}
\value{
List with extended wake change test results
}
\description{
Functions for extended wake change testing, occupancy testing, and other specialized analyses.
Test extended wake change
}
\details{
Tests for changes during extended wake periods across experiments.
}
