% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homeostasis_testing.R
\name{test_homeostasis}
\alias{test_homeostasis}
\title{Homeostasis testing functions}
\usage{
test_homeostasis(response_variable)
}
\arguments{
\item{response_variable}{Name of the response variable to test}
}
\value{
List with homeostasis test results
}
\description{
Functions specific to testing homeostasis hypotheses.
Test homeostasis hypothesis
}
\details{
Tests for homeostasis effects across sleep/wake conditions and experiments.
}
