% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_tests.R
\name{test_interaction_with_experiment}
\alias{test_interaction_with_experiment}
\title{Test interaction with experiment}
\usage{
test_interaction_with_experiment(
  dat,
  models,
  novelty_matrix,
  locomotion_matrix,
  dual_matrix
)
}
\arguments{
\item{dat}{Data frame}

\item{models}{List of nested models (full, reduced, null)}

\item{novelty_matrix}{Contrast matrix for novelty experiment}

\item{locomotion_matrix}{Contrast matrix for locomotion experiment}

\item{dual_matrix}{Contrast matrix for dual experiment}
}
\value{
List with interaction test results
}
\description{
Tests for an interaction between any categorical primary covariate
(e.g., condition, or state) and experiment (as a secondary covariate).
You can also test for interactions with experiment using the more generic
test_interaction function above, but this function is more convenient
for the common case where you have a contrast matrix for each experiment.
}
