% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_testing.R
\name{test_single_day_occupancy}
\alias{test_single_day_occupancy}
\title{Test single day occupancy}
\usage{
test_single_day_occupancy(day)
}
\arguments{
\item{day}{Day to test ("Baseline" or "Recovery")}
}
\value{
List with occupancy test results
}
\description{
Tests occupancy differences between experiments for a single day.
}
