% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_testing.R
\name{test_single_experiment_occupancy}
\alias{test_single_experiment_occupancy}
\title{Test single experiment occupancy}
\usage{
test_single_experiment_occupancy(experiment)
}
\arguments{
\item{experiment}{Experiment to test ("Novelty", "Locomotion", or "Dual")}
}
\value{
List with occupancy test results
}
\description{
Tests occupancy differences between days for a single experiment.
}
